Combine the tables from the ESAS Data Model into a single data.frame

The Analysis functions in this R package expect a single data.frame as input. This function combines the tables from the ESAS Data Model into a single data.frame. It performs left joins to ensure that all records from the CAMPAIGNS, SAMPLES, POSITIONS, and OBSERVATIONS tables are included. It also renames the Notes columns from the CAMPAIGNS and SAMPLES tables to avoid naming conflicts.

Usage

Create_ESAS_Table(esas_tables_list)

Arguments

esas_tables_list

A list containing the ESAS tables as data.frames. The list should include the following named elements: CAMPAIGNS, SAMPLES, POSITIONS, and OBSERVATIONS.

Value

A single data.frame that combines the information from the CAMPAIGNS, SAMPLES, POSITIONS, and OBSERVATIONS tables.

See also

Examples

if (FALSE) { # \dontrun{
# Read 4 ESAS tables:
ESAS_TABLES_LIST <- Read_ESAS_Tables(
  path = "./Data/ESAS download 2024 04 30",
  file_encoding = "UTF-8"
)

# Create an ESAS master-table:
ESAS_TABLE <- Create_ESAS_Table(esas_tables_list = ESAS_TABLES_LIST)
} # }